rename utils to type-utils
