+++ src/package.json
@@ -21,9 +21,9 @@
     "angular4"
   ],
   "peerDependencies": {
-    "@angular/common": ">=4.0.0 <6.0.0",
-    "@angular/core": ">=4.0.0 <6.0.0",
-    "@angular/forms": ">=4.0.0 <6.0.0"
+    "@angular/common": ">=5.0.0 <6.0.0",
+    "@angular/core": ">=5.0.0 <6.0.0",
+    "@angular/forms": ">=5.0.0 <6.0.0"
   },
   "ngPackage": {
     "lib": {
